% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_to_settle_mcp.R
\name{days_to_settle_mcp}
\alias{days_to_settle_mcp}
\title{Calculate Days to Settle Using Minimum Convex Polygon (MCP)}
\usage{
days_to_settle_mcp(x, coords, ID, DateTime, crs)
}
\arguments{
\item{x}{A data frame containing tracking data.}

\item{coords}{A character vector specifying the columns representing the coordinates (e.g., c("UTM_X", "UTM_Y")).}

\item{ID}{A character string specifying the column containing unique animal IDs.}

\item{DateTime}{The column containing datetime information. This must be in standardized format class "POSIXct" "POSIXt"}

\item{crs}{An object specifying the coordinate reference system (e.g., "EPSG:32648"). Note all coordinates should be projected in UTM format for metric measurement, not WGS84 latlong.}
}
\value{
A list containing three elements:
  \itemize{
    \item \code{mcp_all}: A data frame with columns \code{ID}, \code{DateTime}, \code{mcp_area}, and \code{Days} representing the tracked animals, datetime, cumulative MCP area, and number of days, respectively.
    \item \code{summary_nls_res}: A summary of the logistic growth model fit using \code{nls} with details on the coefficients.
    \item \code{nls_plot}: A \code{ggplot} object visualizing the logistic growth model prediction with vertical and horizontal lines indicating the estimated plateau.
  }
}
\description{
The \code{days_to_settle_mcp} function identifies the average number of days it takes for a series of tracked animals to settle into their new environment after translocation or reintroduction. The settling date is determined by updating a convex hull iteratively for each animal ID until the animal only moves within the occupied space, and no longer outside of it, indicating the animal is no-longer going on long-distance exploratory jaunts. A logistic growth model is then fitted using the number of days and updated geometry to simulate a logistic growth curve. Settlement is considered to be achieved when the curve reaches 95% of the plateau.
}
\examples{
\dontrun{
# Example usage
dat <- read.csv("tracking_data.csv")
result <- days_to_settle_mcp(dat, coords = c("UTM_X", "UTM_Y"), ID = "ID", DateTime = datetime, crs = "EPSG:32648")

print(result[[1]])  # sf data frame of the updated convex hull area and days for each animal ID
print(result[[2]])  # summary of the nls model
print(result[[3]])  # ggplot object showing the logistic growth curve results
}

}
\seealso{
\code{\link{st_as_sf}}, \code{\link{st_crs}}, \code{\link{st_distance}}, \code{\link{st_combine}}, \code{\link{st_convex_hull}}, \code{\link{ggplot}}, \code{\link{nls}}
}
