% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_dist.R
\name{disp_dist}
\alias{disp_dist}
\title{Calculate Dispersal Distances}
\usage{
disp_dist(x, coords, ID, DateTime, crs)
}
\arguments{
\item{x}{A data frame containing tracking data.}

\item{coords}{A character vector specifying the columns representing the coordinates (e.g., c("UTM_X", "UTM_Y")).}

\item{ID}{A character string specifying the column containing unique animal IDs.}

\item{DateTime}{The column containing datetime information. This must be in standardized format class "POSIXct" "POSIXt"}

\item{crs}{An object specifying the coordinate reference system (e.g., "EPSG:32648"). Note all coordinates should be projected in UTM format for metric measurement, not WGS84 latlong.}
}
\value{
A data frame with columns \code{ID} and \code{disp_dist} representing animal IDs and dispersal distances, respectively.
}
\description{
The \code{disp_dist} function calculates dispersal distances for wildlife tracking data. You can combine the output via ID with any existing species metadata - e.g. sex, mass, age class, etc. - to identify trends and variations in dispersal distances. This can help in planning, for example, for:
-species with territorial males/females and how far - by averaging the output for each sex - they should be released away from one another.
-to identify dispersal trends in habituated vs. non-habituated animals to inform captive enrichment strategy changes
-to identify dispersal trends between the same species relocated in various sites/ecosystems to see if there are differences that may indicate preference - reduced dispersal could indicate highly abundant resources and lack of need for long-distance exploration.
}
\examples{
\dontrun{
# Example usage
dat <- read.csv("tracking_data.csv")
meta <- read.csv("meta_data.csv")

disp_distances <- disp_dist(dat, coords = c("UTM_X", "UTM_Y"), ID = "ID", DateTime = datetime, crs = "EPSG:32648")

meta <- left_join(meta, disp_distances, by="ID") #note: this uses dplyr function left_join
}

}
\seealso{
\code{\link{st_distance}}, \code{\link{st_as_sf}}, \code{\link{st_crs}}
}
